preprocessor: implement if/else, cleanup code
if-eval code token from my tinycpp project. when i realized i was
basically reimplemting a proper tokenizer and parser, i figured
it would be less error-prone and work to just take my existing
already-debugged code.

note that currently the if-nesting level logic is broken (it was
already), i'll fix it with the logic of tinycpp in the next commit.