preprocessor: add support for #(el)if defined(MACRO)
with this added, the preprocessor finally supports almost the
entire set of conditional functionality of the C cpp.

still lacking are function-like macros, stringification, and
recursive expansion, though i won't implement those, as
ascc can already call out to the real C preprocessor.

i added the conditions functionality because i need them
for agsdefns.sh, which needs to be included for *every* ags
project - so the choice was either to remove the ags
preprocessor completely, or add the required functionality.

note that the ags preprocessor supports the constructs
"ifver", "ifnver", "sectionstart" and "sectionend" - the
latter 2 are nops but exist in almost every source file,
the former 2 might be needed by some stuff and are relatively
awkward to shoehorn into an existing full-blown preprocessor
solution.