agsdefns.sh: some changes to make API version 261 generate correct code
tested with room55.asc extracted with agsutils from Larry 2 Point and Click[0].
basically AGS < 2.71 doesn't implement any pointers, and compiling with
the wrong definitions causes the compiler to emit code for accessing
managed objects, for which the old engine doesn't know the bytecode yet.
this should make it possible to re-inject compiled source into the original
exe and still have it work on windows/wine.
note that there are probably dozens of other cases that need to be fixed,
which i didn't encounter yet.

[0]: https://github.com/rofl0r/agsutils/files/5794186/Larry.2.zip