re-enable (int/short/long)* casts to int with -fpointerhack=1
unfortunately i had to rebase the codebase because of a really
stupid bug introduced in the very first commit that introduced
char* support, which accidentally made it so *any* typechecking
was disabled, and which only fixed very recently.
therefore git bisect couldn't be used at all, and i had to
move the fix way back in time and squash it with the commit
that introduced the bug.
this however results in that later commits which depended on the
broken behaviour no longer get past the typechecking.
this commit restores the no-type-check functionality, so all tests
depending on e.g. int* to int casts work again.