fix line number reporting
until now, all linenumbers were doubled by adding an additional
\r\n pair after every line ending with \n, regardless of whether
it already contained an endline marker.
this occured due to a bug in fmem_gets, which intended to strip
the CRLF or LF from the end of a string, but did so only for CRLF,
whereas fmem_puts assumed to get an input string without either.
my previous attempt to fix only fmem_puts caused the linenumbers
to be almost correct, except that empty lines following a non-empty
line caused the second \n to be swallowed.

improved performance of fmem_puts by using a stack buffer as long
as the input fits in there.