get rid of platform.h
the endianness-specific code was moved to the TU using it,
and all unused code removed - that includes routines changing
big endian input bytes to host order as well as floating point
swapping.
this allows us to use the portable fallback code of endianness.h
in case the compile-time detection of the endianness fails.